% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{processLogiOper}
\alias{processLogiOper}
\title{Processing Logical Operations}
\usage{
processLogiOper(logicalOperation, stringList)
}
\arguments{
\item{logicalOperation}{A String containing the logical operation to be performed}

\item{stringList}{A list of String over which the logical operation is to be performed}
}
\value{
A list of the output values(0/1)
}
\description{
Dynamically process the logical operations over a list of string and get the result
}
\examples{
logicalOperation <- "AND(Cake,OR(Birthday,Anniversary))"
stringList <- c("The cake at the birthday was awesome."
                ,"Their anniversary was last week.")

processLogiOper(logicalOperation, stringList)
}

